Moved ParagraphNode into a separate file;
