+++ packages/flame/lib/src/text/nodes/document_node.dart
@@ -4,6 +4,7 @@ import 'package:flame/src/text/common/utils.dart';
 import 'package:flame/src/text/elements/element.dart';
 import 'package:flame/src/text/elements/group_element.dart';
 import 'package:flame/src/text/nodes.dart';
+import 'package:flame/src/text/nodes/paragraph_node.dart';
 import 'package:flame/src/text/styles/document_style.dart';
 import 'package:flutter/painting.dart';
