+++ packages/flame/lib/src/text/styles/document_style.dart
@@ -1,4 +1,5 @@
 import 'package:flame/src/text/nodes.dart';
+import 'package:flame/src/text/nodes/paragraph_node.dart';
 import 'package:flame/src/text/styles/background_style.dart';
 import 'package:flame/src/text/styles/block_style.dart';
 import 'package:flame/src/text/styles/overflow.dart';
